+++ src/components/Navbar.jsx
@@ -39,6 +39,9 @@ const Navbar = () => {
 
 const Nav = styled.nav`
   border-bottom: 1px solid ${COLORS.gray200};
+  position: relative;
+  z-index: 9999;
+  background-color: ${COLORS.white};
 `;
 
 const Container = styled.div`
